exception handling delete

+++ src/main/java/com/wisan/project/services/UserService.java
@@ -3,6 +3,7 @@ package com.wisan.project.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.wisan.project.entities.User;
@@ -22,16 +23,25 @@ public class UserService {
 	public User findById(Long id) {
 		return userRepository.findById(id).orElseThrow(() -> new ResourceNotFundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return userRepository.saveAndFlush(obj);
-		
+
 	}
-	
+
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if (userRepository.existsById(id)) {
+				userRepository.deleteById(id);
+			} else {
+				throw new ResourceNotFundException(id);
+			}
+		} catch (DataIntegrityViolationException e) {
+			throw new DataIntegrityViolationException("Data integrity violation occurred.");
+
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = userRepository.getReferenceById(id);
 		updateData(entity, obj);
@@ -42,7 +52,7 @@ public class UserService {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
-		
+
 	}
 
 }
